requirements-dev.txt
ruff
requests_cache
responses
pytest
pytest-env
